+++ docs/changelog.rst
@@ -28,3 +28,10 @@ Additions to the :doc:`supporting modules <supporting-modules>`.
     * :py:func:`salespyforce.utils.tests.resources.instantiate_with_local_helper`
 * Added the :py:mod:`salespyforce.utils.tests.test_instantiate_object` module with the following functions:
     * :py:func:`salespyforce.utils.tests.test_instantiate_object.test_instantiate_core_object`
+    * :py:func:`salespyforce.utils.tests.test_instantiate_object.test_get_api_versions`
+    * :py:func:`salespyforce.utils.tests.test_instantiate_object.test_get_rest_resources`
+* Added the :py:mod:`salespyforce.utils.tests.test_sobjects` module with the following functions:
+    * :py:func:`salespyforce.utils.tests.test_sobjects.test_get_all_sobjects`
+    * :py:func:`salespyforce.utils.tests.test_sobjects.test_get_and_describe_sobject`
+* Added the :py:mod:`salespyforce.utils.tests.test_soql` module with the following functions:
+    * :py:func:`salespyforce.utils.tests.test_soql.test_soql_query`