+++ docs/changelog.rst
@@ -6,7 +6,7 @@ This page documents the additions, changes, fixes, deprecations and removals mad
 ******
 v1.1.0
 ******
-**Release Date: TBD**
+**Release Date: 2023-05-29**
 
 Added
 =====